move some swing methods to SwingUtils, fix menus not getting destroyed properly
