+++ src/main/java/amidst/gui/export/BiomeExporterDialog.java
@@ -53,7 +53,6 @@ import javax.swing.SwingConstants;
 import javax.swing.SwingUtilities;
 import javax.swing.border.LineBorder;
 
-import amidst.Amidst;
 import amidst.documentation.NotThreadSafe;
 import amidst.gui.main.Actions;
 import amidst.gui.main.PNGFileFilter;
@@ -71,6 +70,7 @@ import amidst.mojangapi.world.coordinates.CoordinatesInWorld;
 import amidst.mojangapi.world.oracle.BiomeDataOracle;
 import amidst.settings.Setting;
 import amidst.settings.biomeprofile.BiomeProfileSelection;
+import amidst.util.SwingUtils;
 
 @NotThreadSafe
 public class BiomeExporterDialog {
@@ -423,7 +423,7 @@ public class BiomeExporterDialog {
 
 	public void dispose() {
 		menuBarSupplier.get().setMenuItemsEnabled(new String[] { "Export Biomes to Image ...", "Biome Profile" }, true);
-		Amidst.destroyComponentTree(dialog);
+		SwingUtils.destroyComponentTree(dialog);
 	}
 	
 	public void softDispose() {
